pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "EzSQL.py"
version = "0.3.0"
description = "A small, beginner-friendly SQL wrapper for Python DB-API drivers."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "srinjan93306-ai" }
]
keywords = ["sql", "sqlite", "postgres", "mysql", "oracle", "database", "db-api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
]
dependencies = []

[project.optional-dependencies]
postgres = ["psycopg2>=2.9"]
mysql = ["mysql-connector-python>=8.0"]
oracle = ["oracledb>=2.0"]
all = [
    "psycopg2>=2.9",
    "mysql-connector-python>=8.0",
    "oracledb>=2.0",
]

[project.urls]
Homepage = "https://github.com/srinjan93306-ai/ezsql"
Issues = "https://github.com/srinjan93306-ai/ezsql/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ezsql*"]

[tool.setuptools]
py-modules = ["EzSQL"]

[tool.setuptools.package-data]
ezsql = ["py.typed"]
